added NearTypes, renamed ReactChildrenType.js to ReactChildrenTypes.js
