+++ src/router/GuardedRoute.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
-import { ReactChildrenType } from '~types/ReactChildrenType';
+import { ReactChildrenType } from '~types/ReactChildrenTypes';
 
 const GuardedRoutePropTypes = {
   children: ReactChildrenType,
