+++ src/types/NearTypes/NearConfigType.js
@@ -1,0 +1,5 @@
+import PropTypes from 'prop-types';
+
+export const NearConfigType = { contractName: PropTypes.string.isRequired };
+
+export const NearConfigTypeShape = PropTypes.shape(NearConfigType);
