+++ src/types/NearTypes/NearType.js
@@ -1,0 +1,5 @@
+import PropTypes from 'prop-types';
+
+export const NearType = { connection: PropTypes.object.isRequired };
+
+export const NearTypeShape = PropTypes.shape(NearType);
