+++ src/types/NearTypes/UserTypes.js
@@ -1,0 +1,5 @@
+import PropTypes from 'prop-types';
+
+export const UserType = { accountId: PropTypes.string.isRequired, balance: PropTypes.string.isRequired };
+
+export const UserTypeShape = PropTypes.shape(UserType);
