+++ src/types/NearTypes/index.js
@@ -1,0 +1,15 @@
+import { NearConfigType, NearConfigTypeShape } from './NearConfigType';
+import { NearType, NearTypeShape } from './NearType';
+import { UserTypes, UserTypeShape } from './UserTypes';
+import { WalletType, WalletTypeShape } from './WalletType';
+
+export {
+  NearConfigType,
+  NearConfigTypeShape,
+  NearType,
+  NearTypeShape,
+  UserTypes,
+  UserTypeShape,
+  WalletType,
+  WalletTypeShape,
+};
